Renamed HomeController to E2EController to ease copying
