+++ E2EDotNet.Tests/E2EDotNet.Tests.csproj
@@ -111,7 +111,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Controllers\HomeControllerTest.cs" />
+    <Compile Include="Controllers\E2EControllerTest.cs" />
     <Compile Include="TestRunnerTests.cs" />
   </ItemGroup>
   <ItemGroup>
